requirements.txt
requests==2.25.1
discord-webhook==0.12.0
PyYAML==5.4.1

